Last version of Groupmanagement

+++ Connect_Hub/src/Backend/GroupManagement/PrimaryAdminRole.java
@@ -1,12 +1,14 @@
 package Backend.GroupManagement;
 
 import Backend.User.User;
-
+import Backend.Database.*;
 public class PrimaryAdminRole extends AdminRole {
 
     private static volatile PrimaryAdminRole instance;
 
+
     private PrimaryAdminRole() {}
+
 
     // Public method to get the singleton instance
     public static PrimaryAdminRole getInstance() {
@@ -23,22 +25,28 @@ public class PrimaryAdminRole extends AdminRole {
     public void removeMember(Group group, User user) {
         if (group.isMember(user)) {
             group.removeMember(user);
+            refresh();
         }
+
     }
 
     public void promoteToAdmin(Group group, User user) {
         group.promoteToAdmin(user);
+        refresh();
     }
 
     public void demoteAdmin(Group group, User user) {
         if (group.isAdmin(user)) {
             group.demoteAdmin(user);
+            refresh();
         }
+
     }
 
     public void deleteGroup(User user, Group group) {
-        if (group.isPrimaryAdmin(user)) {
+        if (groupsDataBase.IsGroupFound(group)&&group.isPrimaryAdmin(user)) {
             user.getGroupManager().deleteGroup(group);
+            groupsDataBase.removeGroup(group);
         }
     }
 }
